fix: calculate SMP for very near skeletons even if their origin point is hidden
We now always enable the skeletons that are just around the camera.
It's useful if for example the skeleton origin is very near, behind the camera,
but some parts or the skeleton are in front of the camera and need to be
animated.

This is implemented with a minCullingDistance parameter: any skeleton
nearer than this distance is always calculated. Default value is 500.

Can be disabled by setting the distance to zero.